Add inherit tests for windows, linux and darwin
This is neccessary because windows uses %VAR% for env variable
expansions, on the other side linux and darwin both are using the unix
standard which uses $VAR.
